Make a transitions module file
