--- Transisions.js
@@ -1,6 +1,0 @@
-const contact = () => {
-   body.innerHTML = "h"
-    
-  }
-
-  export {contact}
