+++ index.html
@@ -6,6 +6,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Awsome books</title>
     <link rel="stylesheet" href="index.css">
+    <script type="module" src="index.js"></script>
 </head>
 <body>
     <nav><div class="awesome-nav">
@@ -54,6 +55,6 @@
     </ul>
 </div>
 <footer>Copyright ...</footer>
-   <script type="module" src="index.js"></script>
+
 </body>
 </html>
